set value test
