+++ src/Action/SetValue.php
@@ -4,20 +4,28 @@
 namespace Debuqer\TikaFormBuilder\Action;
 
 
+use Debuqer\TikaFormBuilder\Exceptions\NotPropertySettingSupport;
 use Debuqer\TikaFormBuilder\Form;
 
 class SetValue extends BaseAction
 {
     public function run(Form &$form)
     {
-        $fieldName = $this->getParameters()->get('field');
-        $attribute = $this->getParameters()->get('attribute');
+        $fieldAddress = $this->getParameters()->get('field');
+        $address = explode('.', $fieldAddress);
+        $fieldName = implode('.', array_slice($address, 0, sizeof($address) - 1));
+        $fieldAttribute = $address[sizeof($address) - 1];
 
         $expr = $this->getParameters()->get('value');
         $value = $this->expressionLanguage->evaluate($expr, [
             'form' => $form,
         ]);
 
-        $form->getInstance()->getItems()->get($fieldName)->setProperty($attribute, $value);
+        $item = $form->get($fieldName);
+        if( method_exists($item, 'setProperty') ) {
+            $form->get($fieldName)->setProperty($fieldAttribute, $value);
+        } else {
+            throw new NotPropertySettingSupport(sprintf('Item %s does not have setProperty method', $fieldName));
+        }
     }
 }
