+++ src/Exceptions/NotPropertySettingSupport.php
@@ -1,0 +1,7 @@
+<?php
+namespace Debuqer\TikaFormBuilder\Exceptions;
+
+class NotPropertySettingSupport extends \Exception
+{
+
+}
